+++ lib/linters/example-linter.js
@@ -2,7 +2,7 @@ const fs = require('fs')
 const path = require('path')
 const Ajv = require('ajv')
 const yaml = require('js-yaml')
-const pluginYamlParser = require('../plugin-yaml-parser')
+const pluginYamlParser = require('../plugin-yaml')
 
 module.exports = function(pluginName, pluginPath, options) {
   const readmePath = path.join(pluginPath, options.readme)
