+++ lib/linters/plugin-yaml-linter.js
@@ -1,13 +1,17 @@
-const pluginYamlParser = require('../plugin-yaml-parser')
+const path = require('path')
+const fs = require('fs')
+const pluginYamlParser = require('../plugin-yaml')
+const yaml = require('js-yaml')
 const Ajv = require('ajv')
-
-const schema = require("../plugin-yaml-schema.json")
 
 module.exports = function(pluginName, pluginPath, options) {
   const pluginYaml = pluginYamlParser(pluginPath)
   const silent = options.silent
 
   const ajv = new Ajv({ allErrors: true, jsonPointers: true })
+
+  const schema = yaml.safeLoad(fs.readFileSync(path.join(__dirname, '..', 'plugin-yaml-schema.yml'), 'utf8'))
+
   const validator = ajv.compile(schema)
 
   const valid = validator(pluginYaml)
