+++ lib/plugin-yaml-schema.yml
