+++ lib/plugin-yaml.js
