+++ test/example-linter.test.js
@@ -1,12 +1,12 @@
 const assert = require('chai').assert
-const lintExamples = require('../lib/linters/lint-examples')
+const exampleLinter = require('../lib/linters/example-linter')
 const path = require('path')
-const fixtures = path.join(__dirname, 'lint-examples')
+const fixtures = path.join(__dirname, 'example-linter')
 
 describe('lint-examples', () => {
   describe('valid example', () => {
     it('should be valid', () => {
-      assert(lintExamples('valid-plugin', path.join(fixtures, 'valid-plugin'), {
+      assert(exampleLinter('valid-plugin', path.join(fixtures, 'valid-plugin'), {
         silent: true,
         readme: 'README.md'
       }))
@@ -14,7 +14,7 @@ describe('lint-examples', () => {
   })
   describe('valid plugin with yaml instead of yml', () => {
     it('should be valid', () => {
-      assert(lintExamples('valid-plugin-with-yaml', path.join(fixtures, 'valid-plugin-with-yaml'), {
+      assert(exampleLinter('valid-plugin-with-yaml', path.join(fixtures, 'valid-plugin-with-yaml'), {
         silent: true,
         readme: 'README.md'
       }))
@@ -22,7 +22,7 @@ describe('lint-examples', () => {
   })
   describe('invalid examples', () => {
     it('should be invalid', () => {
-      assert(!lintExamples('invalid-examples', path.join(fixtures, 'invalid-examples'), {
+      assert(!exampleLinter('invalid-examples', path.join(fixtures, 'invalid-examples'), {
         silent: true,
         readme: 'README.md'
       }))
@@ -30,7 +30,7 @@ describe('lint-examples', () => {
   })
   describe('custom readme paths', () => {
     it('should work', () => {
-      assert(lintExamples('custom-readme', path.join(fixtures, 'custom-readme'), {
+      assert(exampleLinter('custom-readme', path.join(fixtures, 'custom-readme'), {
         silent: true,
         readme: 'custom-readme.md'
       }))
@@ -38,7 +38,7 @@ describe('lint-examples', () => {
   })
   describe('skips validation if missing .schema', () => {
     it('should work', () => {
-      assert(lintExamples('missing-schema', path.join(fixtures, 'missing-schema'), {
+      assert(exampleLinter('missing-schema', path.join(fixtures, 'missing-schema'), {
         silent: true,
         readme: 'README.md'
       }))
