+++ test/example-linter/custom-readme/plugin.yml
