+++ test/example-linter/invalid-examples/plugin.yml
