+++ test/example-linter/missing-schema/plugin.yml
