+++ test/example-linter/valid-plugin-with-yaml/README.md
