+++ test/example-linter/valid-plugin-with-yaml/plugin.yml
