+++ test/example-linter/valid-plugin/plugin.yml
