+++ test/plugin-yaml-linter/valid-plugin/plugin.yml
